+++ src/layouts/AppLayout.tsx
@@ -5,8 +5,8 @@ import { Outlet, useLocation } from 'react-router-dom';
 import 'react-toastify/dist/ReactToastify.css';
 
 // components
-import Sidebar from '../components/global/Sidebar';
-import Appbar from '../components/global/Appbar';
+import Sidebar from '../components/admin/Sidebar';
+import Appbar from '../components/admin/Appbar';
 
 const AppLayout = ({ isAdmin }: { isAdmin: boolean }) => {
   const location = useLocation();
