+++ src/pages/admin/Dashboard.tsx
@@ -1,9 +1,7 @@
 import React from 'react';
 
 const Dashboard = () => {
-    return (
-        <div className='text-darkColor dark:text-lightColor'>Dashboard</div>
-    )
-}
+  return <div className='text-darkColor dark:text-lightColor'>Dashboard</div>;
+};
 
-export default Dashboard
+export default Dashboard;
